Set 'CscToolPath' to OmniSharp's local MSBuild on Mono
