+++ src/OmniSharp.Host/MSBuild/Discovery/Providers/StandAloneInstanceProvider.cs
@@ -18,7 +18,7 @@ namespace OmniSharp.MSBuild.Discovery.Providers
 
         public override ImmutableArray<MSBuildInstance> GetInstances()
         {
-            var path = FindMSBuildDirectory();
+            var path = FindLocalMSBuildDirectory();
             if (path == null)
             {
                 return ImmutableArray<MSBuildInstance>.Empty;
@@ -59,42 +59,6 @@ namespace OmniSharp.MSBuild.Discovery.Providers
                     DiscoveryType.StandAlone,
                     propertyOverrides.ToImmutable(),
                     setMSBuildExePathVariable: true));
-        }
-
-        private static string FindMSBuildDirectory()
-        {
-            // If OmniSharp is running normally, MSBuild is located in an 'msbuild' folder beneath OmniSharp.exe.
-            var msbuildDirectory = Path.Combine(AppContext.BaseDirectory, "msbuild");
-
-            if (!Directory.Exists(msbuildDirectory))
-            {
-                // If the 'msbuild' folder does not exists beneath "OmniSharp.exe", this is likely a development scenario,
-                // such as debugging or running unit tests. In that case, we try to locate the ".msbuild" folder at the
-                // solution level.
-                msbuildDirectory = FindLocalMSBuildFromSolution();
-            }
-
-            return msbuildDirectory;
-        }
-
-        private static string FindLocalMSBuildFromSolution()
-        {
-            // Try to locate the .msbuild folder by search for the OmniSharp solution relative to the current folder.
-            var current = AppContext.BaseDirectory;
-            while (!File.Exists(Path.Combine(current, "OmniSharp.sln")))
-            {
-                current = Path.GetDirectoryName(current);
-                if (Path.GetPathRoot(current) == current)
-                {
-                    break;
-                }
-            }
-
-            var result = Path.Combine(current, ".msbuild");
-
-            return Directory.Exists(result)
-                ? result
-                : null;
         }
 
         private static bool TryGetMonoXBuildPath(out string path)
